requirements.txt
fastapi
uvicorn[standard]
httpx
openai
anthropic
langchain-anthropic
langgraph
fatsecret
sqlalchemy
psycopg2-binary
pydantic-settings
apscheduler
python-dotenv
